Remove ImageGrid component
